fix(security): stabilize delegated authorization
Extract portable import authorization into a focused service, preserve existing tenant constraint behavior for medication locations, and make dashboard fallback person selection deterministic.
